requirements.txt
Flask==2.0.2
numpy==1.23.1
pandas==1.4.3
Pillow==9.3.0
psycopg2==2.9.5
psycopg2_binary==2.9.3
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.41
st_clickable_images==0.0.3
streamlit==1.15.2
tabula==1.0.5
tabula_py==2.6.0
